==> scripts/01. deck만들기.ipynb <==
%load_ext lab_black
%cd ..

import argparse
import math
import multiprocessing
import sys
from datetime import datetime
from pathlib import Path
from pprint import pformat

import numpy as np
import pandas as pd
import torch
import torch.nn as nn
import torch.nn.functional as F
import torch_optimizer
import yaml
from easydict import EasyDict
from pytorch_transformers import DistilBertForSequenceClassification, DistilBertTokenizer

from datasets.dataset_ver7 import DatasetGeneratorVer7
from utils import SAM, AverageMeter, CustomLogger, FocalLoss, seed_everything

from main_ver7 import MyTrainer
from collections import defaultdict
import matplotlib.pyplot as plt
import random
from sklearn.metrics import f1_score, classification_report
from tqdm import tqdm

postfix = "distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_1"
outdir = Path("results/distilbert-base-uncased-ver7")
fold = 1

with open("config/distilbert-base-uncased-ver7.yaml", "r") as f:
    C = EasyDict(yaml.load(f, yaml.FullLoader))
    C.result_dir = Path(C.result_dir)
    C.dataset.dir = Path(C.dataset.dir)
    seed_everything(C.seed, deterministic=False)

C

trainer = MyTrainer(C, fold, outdir / f"{postfix}.pth")

model = trainer.model
model.eval()
torch.set_grad_enabled(False)

activation = []


def hook(model, input, output):
    activation.append(output.detach().cpu())

model.pre_classifier.register_forward_hook(hook)

train_dl = trainer.dsgen.train_only()

# 덱만들기 (Train)

activation = []
deck = {"feat": [], "otext": [], "tlevel": [], "fclevel": []}
with tqdm(total=len(train_dl.dataset), ncols=100, file=sys.stdout) as t:
    for text, tlevel, otext in train_dl:
        fclevel = model(text.cuda(non_blocking=True))[0].argmax(dim=1).cpu()
        deck["fclevel"].append(fclevel)
        deck["tlevel"].append(tlevel)
        deck["otext"].extend(otext)
        t.update(text.size(0))

deck["feat"] = torch.cat(activation)
deck["tlevel"] = torch.cat(deck["tlevel"])
deck["fclevel"] = torch.cat(deck["fclevel"])

deck["tlevel"].shape, deck["fclevel"].shape

deck["feat"].shape

torch.save(deck, outdir / f"{postfix}-deck1.pth")

# Valid Lv7에 대한 결과 분석

deck = torch.load("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_1.pth")

ds_lv7 = trainer.dsgen.valid_lv7().dataset

text, otext = ds_lv7[0]
activation = []
print(model(text[None].cuda())[0].cpu())

dists, indices = torch.norm(deck["feat"] - activation[0][None], p=None, dim=1).topk(4, largest=False)
dists, indices, deck["tlevel"][indices]

text, otext = ds_lv7[1]
activation = []
print(model(text[None].cuda())[0].cpu())

dists, indices = torch.norm(deck["feat"] - activation[0][None], p=None, dim=1).topk(4, largest=False)
dists, indices, deck["tlevel"][indices]

text, otext = ds_lv7[2]
activation = []
print(model(text[None].cuda())[0].cpu())

dists, indices = torch.norm(deck["feat"] - activation[0][None], p=None, dim=1).topk(4, largest=False)
dists, indices, deck["tlevel"][indices]

dists, indices = torch.norm(deck["feat"] - activation[0][None], p=None, dim=1).topk(4, largest=False)
dists, indices, deck["tlevel"][indices]

# 덱만들기 (Submission)

ds_test = trainer.dsgen.test()

activation = []
deck = {"feat": [], "otext": [], "fclevel": [], "ids": []}
with tqdm(total=len(ds_test), ncols=100, file=sys.stdout) as t:
    for i in range(len(ds_test)):
        text, otext, ids = ds_test[i]
        fclevel = model(text[None].cuda(non_blocking=True))[0].argmax(dim=1).cpu()
        deck["fclevel"].append(fclevel)
        deck["otext"].append(otext)
        deck["ids"].append(ids)
        t.update()

deck["feat"] = torch.stack(activation)
deck["fclevel"] = torch.stack(deck["fclevel"])

deck["feat"].shape, deck["fclevel"].shape

deck["feat"] = deck["feat"][:, 0, :]
# deck["fclevel"] = deck["fclevel"][:, 0]

deck["feat"].shape, deck["fclevel"].shape

torch.save(deck, outdir / f"{postfix}-deck2.pth")
