==> scripts/04. ensemble.ipynb <==
%load_ext lab_black
%cd ..

import numpy as np
import pandas as pd

csvs = [
    pd.read_csv("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_1-ver4.csv"),
    pd.read_csv("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_2-ver4.csv"),
    pd.read_csv("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_3-ver4.csv"),
    pd.read_csv("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_4-ver4.csv"),
    pd.read_csv("results/distilbert-base-uncased-ver7/distilbert-base-uncased-focal-AdamW-lr1e-05-ver7-os10_5-ver4.csv"),
]

levels = np.stack([csv.level.to_numpy() for csv in csvs])

N = len(csvs[0])

out_levels = []
for i in range(N):
    kt = False
    ll = [0 for _ in range(8)]
    for j in range(len(csvs)):
        if levels[j, i] in [2, 4, 6]:
            out_levels.append(levels[j, i])
            kt = True
            break
        else:
            ll[levels[j, i]] += 1

    if not kt:
        if max(ll) == 1:
            out_levels.append(levels[0, i])
        elif ll[7] >= 2:
            out_levels.append(7)
        else:
            out_levels.append(max(range(8), key=lambda i: ll[i]))

df = pd.DataFrame({"id": csvs[0].id.to_list(), "level": out_levels})

df

df.to_csv("results/ens-ver5.csv", index=False)
